lamusitheque/requirements.txt
alabaster==0.7.12
amqp==2.3.2
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.6.0
billiard==3.5.0.4
cachetools==2.1.0
celery==4.2.1
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
decorator==4.3.2
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-appconf==1.0.2
django-braces==1.13.0
django-comments-xtd==2.2.0
django-compressor==2.2
django-contrib-comments==1.8.0
django-cors-headers==3.1.1
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-etc==0.10.0
django-extensions==2.1.5
django-filter==2.1.0
django-filters==0.2.1
django-forms-bootstrap==3.1.0
django-htmlmin==0.10.0
django-jchart==0.4.2
django-jsonfield==1.0.1
django-jsonfield-compat==0.4.4
django-libs==2.0.1
django-messages==0.5.3
django-model-utils==3.1.2
django-notifications-hq==1.5.0
django-postman==3.6.1
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-siteflags==0.5.0
django-social-share==1.3.0
django-tag-parser==3.1
django-threadedcomments==1.2
django-unused-media==0.1.12
django-vote==2.1.7
djangorestframework==3.8.2
djangorestframework-filters==1.0.0.dev0
djangorestframework-recursive==0.1.2
docutils==0.14
drf-nested-routers==0.91
drf-rw-serializers==1.0.1
html5lib==1.0.1
httplib2==0.11.3
idna==2.7
imagesize==1.1.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.2
Jinja2==2.10
jsonfield==2.0.2
kombu==4.2.1
MarkupSafe==1.0
numpy==1.18.2
oauthlib==2.1.0
openapi-codec==1.3.2
packaging==18.0
pandas==1.0.3
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.2.0
pinax-badges==2.0.0
pkg-resources==0.0.0
prompt-toolkit==2.0.8
psycopg2==2.7.5
ptyprocess==0.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
Pygments==2.2.0
pyparsing==2.2.2
python-akismet==0.4.1
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.5
rcssmin==1.0.6
requests==2.19.1
requests-oauthlib==1.0.0
rest-framework-generic-relations==1.2.1
rjsmin==1.0.12
rsa==3.4.2
selenium==3.141.0
simplejson==3.16.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.0
sqlparse==0.2.4
swapper==1.1.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==3.3.1
youtube-data-api==0.0.17
